requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
h11==0.14.0
idna==3.4
packaging==23.0
Pillow==9.4.0
pydantic==1.10.7
pytesseract==0.3.10
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
